+++ lib/services/vector_clock_service.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
 import 'package:crypto/crypto.dart';
+import 'package:lotti/database/settings_db.dart';
 import 'package:lotti/get_it.dart';
-import 'package:lotti/sync/secure_storage.dart';
 import 'package:lotti/sync/utils.dart';
 import 'package:lotti/sync/vector_clock.dart';
 import 'package:lotti/utils/file_utils.dart';
@@ -15,8 +15,8 @@ class VectorClockService {
   late String _host;
 
   Future<void> init() async {
-    _host = await getHost() ?? await setNewHost();
-    _nextAvailableCounter = await _getNextAvailableCounter();
+    _host = await _getHost() ?? await setNewHost();
+    await _getNextAvailableCounter();
   }
 
   Future<void> increment() async {
@@ -26,34 +26,41 @@ class VectorClockService {
 
   Future<String> setNewHost() async {
     final host = uuid.v4();
-    await getIt<SecureStorage>().writeValue(hostKey, host);
+
+    await getIt<SettingsDb>().saveSettingsItem(hostKey, host);
+    await setNextAvailableCounter(0);
+
+    _host = host;
     return host;
   }
 
+  Future<String?> _getHost() async {
+    return getIt<SettingsDb>().itemByKey(hostKey);
+  }
+
   Future<String?> getHost() async {
-    return getIt<SecureStorage>().readValue(hostKey);
+    return _host;
   }
 
   Future<void> setNextAvailableCounter(int nextAvailableCounter) async {
-    await getIt<SecureStorage>().writeValue(
+    _nextAvailableCounter = nextAvailableCounter;
+
+    await getIt<SettingsDb>().saveSettingsItem(
       nextAvailableCounterKey,
       nextAvailableCounter.toString(),
     );
-    _nextAvailableCounter = nextAvailableCounter;
   }
 
-  Future<int> _getNextAvailableCounter() async {
-    int? nextAvailableCounter;
-    final nextAvailableCounterString =
-        await getIt<SecureStorage>().readValue(nextAvailableCounterKey);
+  Future<void> _getNextAvailableCounter() async {
+    final existing = await getIt<SettingsDb>()
+        .watchSettingsItemByKey(nextAvailableCounterKey)
+        .first;
 
-    if (nextAvailableCounterString != null) {
-      nextAvailableCounter = int.parse(nextAvailableCounterString);
+    if (existing.isNotEmpty) {
+      _nextAvailableCounter = int.parse(existing.first.value);
     } else {
-      nextAvailableCounter = 0;
-      await setNextAvailableCounter(nextAvailableCounter);
+      await setNextAvailableCounter(0);
     }
-    return nextAvailableCounter;
   }
 
   Future<int> getNextAvailableCounter() async {